+++ Assets/Scripts/EntityComponent/Skill/SkillBase.cs
@@ -10,7 +10,7 @@ namespace Skill
 
         public int MasterDataId { get; protected set; } = -1;
 
-        public float CoolTime { get; protected set; }
+        public double CoolTime { get; protected set; }
         public bool IsCoolTime => CoolTime > 0;
 
         protected float m_fLastUpdateTime = -1f;
@@ -18,9 +18,9 @@ namespace Skill
         protected int startTick = -1;
         protected int lastTick = -1;
 
-        protected float DeltaTime => lastTick == -1 ? CurrentUpdateTime : CurrentUpdateTime - LastUpdateTime;
-        protected float CurrentUpdateTime => Game.Current.CurrentTick == 0 ? 0 : (Game.Current.CurrentTick - startTick + 1) * Game.Current.TickInterval;
-        protected float LastUpdateTime => lastTick == -1 ? -1 : (lastTick - startTick + 1) * Game.Current.TickInterval;
+        protected double DeltaTime => lastTick == -1 ? CurrentUpdateTime : CurrentUpdateTime - LastUpdateTime;
+        protected double CurrentUpdateTime => Game.Current.CurrentTick == 0 ? 0 : (Game.Current.CurrentTick - startTick + 1) * Game.Current.TickInterval;
+        protected double LastUpdateTime => lastTick == -1 ? -1 : (lastTick - startTick + 1) * Game.Current.TickInterval;
 
         private MasterData.Skill masterData = null;
         public MasterData.Skill MasterData => masterData ?? (masterData = MasterDataManager.Instance.GetMasterData<MasterData.Skill>(MasterDataId));
