+++ Assets/Scripts/SubGame/JumpWang/MoveController.cs
@@ -44,7 +44,7 @@ namespace JumpWangController
             {
                 var entity = Entities.Get<LOPMonoEntityBase>(Entities.MyEntityID);
 
-                entity.BehaviorController.Move(entity.Position + moveInput.Value.direction.normalized * Game.Current.TickInterval * entity.FactoredMovementSpeed);
+                entity.BehaviorController.Move(entity.Position + moveInput.Value.direction.normalized * (float)Game.Current.TickInterval * entity.FactoredMovementSpeed);
             }
 
             moveInput = null;
